Refactor: move PageIndicator to components package
